+++ travis/__main__.py
@@ -1,3 +1,8 @@
-from travis.encyrpt import cli
+"""Run Travis Encrypt from the command line.
+
+Example: python -m travis
+"""
+
+from travis.encrypt import cli
 
 cli()